Fixed my having accidentally swapped the mnd and int checks.
Also I think at some point I thought I was going to do a params table similar to weaponskills and spells use, but then realized that conflicted with the existing params table being used in the magic resists so I renamed it, but then i scrapped the idea and left "ae" barfed all over. Got rid of that and tabled the functions instead.

Still pending: less if/else insanity...