Add a simple flow to allow changing user passwords.
This is a prerequisite for
DarkstarProject/xiloader#7.

This change adds support for a new command in the login code to support
changing a users password. It simply verifies the users login similarly
to the existing flow, and then makes another request over the socket for
the new password. Then it updates the SQL.

Tested locally. Have a set of changes ready for xiloader as well, but
this work is required first.

NOTE: There is definitely some refactoring that could be done in this
file to reduce the amount of nested braces and to reduce duplicated code.
I tried to touch as little existing code as possible in this change to
reduce the risk of introducing a regression.