Add configurable subjob cap ratio..
..and an option to apply to mobs that I hope is eventually deprecated by the ability to manually set mobs sj lv range.

The player options also apply to pets with the exception of charm pets, since those are mobs.

if `subjob_ratio` is 1, player subjobs are normal (one half) as in retail. This is the default option.
2 will set cap of two-thirds, so a lv 75 player can have a subjob up to lv 50, or a lv 99 player up to lv66
3 will make subjobs capped even to mainjob.
0 will make subjobs lv 0 and grant nothing.

I did not want to do this as a float value divisor or multiplier for several reasons, not the least of which is explaining to people the proper math with get the ration they intended.

`include_mob_sj` determines if the adjustments also apply to monsters.

NOTE: since `SetSLevel()` handled the cap in its entirely, all existing attempts to set the subjob should lead back to here and rely on this. However any instance where a previous editor did not know this probably ***already had incorrect sub levels before I changed anything.***

Todo: review petutils section dealing with pups automaton, because it looks like the auto's subjob is being manually cut in half before it would have been cut in half by the function ***before I even touched anything*** which doesn't seem at all right since that would mean half of a half of a subjob and whats even the point of that.