Added already coded spells to RUN
Updated binary code to allow RUN to cast spells they have. 

Crusade, and Regen IV are not currently coded and not able to update for run specifically. Foil is still commented out and I'm not sure if that is because RUN is technically NYI at this time or because something is wrong with it.