PR: lua: Fixed bug with Violent Flourish
When calling cMeleeRatio, need to pass in TP. This was bugging
out here causing the whole script for violent flourish to fail.

Passed in zero TP, works fine, but the fTP function (which was
the one that originally was failing) expects TP >= 1000. It
prints out a simple error message. Can be ignored, didn't do
further work because this whole system needs a rework, what
an absolute mess. It doesn't need more technical debt by
me preventing this error msg.