This mimics the changes done in battleutils for DexCritRate, Hit Rate, fSTR, pDIF and Level Correction. I have also removed the cap from the summoning skill bonus, it was set to 200 but retail is well over 250 now. My SMN skill set has 695 total skill and I know I can break 700. I have moved target:WakeUp to after the fTP damage modifier instead of before just in case.
As a note Blood Pact damage is still not accurate with this change but this ensures that the basic computations match retail.
It seems fTP values are somewhat incorrect as well as how fTP can cary over between hits which is confirmed on some BPs.
I plan to work towards fixing Blood Pact equations in another PR but I feel like this is large enough and focused on base damage calculations and hit rate.