Finished adjusting the co-ordinates for where the player will be placed after selecting a zone to teleport to thru the Survival Guide teleport system.
The player will be placed facing the guide (as if they were reading it) or to the left of the guide (still facing the guide) if navmesh/environment limitations didn't permit the first option.

Second option was only used on less than a handful of zones.

Placement away from the guide was also standardized to be within a set distance range (around 1.1 to 1.3 yalms away) (range used instead of because of the bobbing of the guides)