Change Nidhogg's "Super Buff" skill ID (#1464)
* Add super_buff at its legit position

* Use legit skillID for super_buff

* Fix description

* do not require global settings

ibm2431's note: This new skill ID was confirmed from a retail capture. It is the true one which Nidhogg uses for its "2hr animation". The previous ID was a "shortcut" which was assumed to never be used by anything, but was later discovered to be used by Wyrm (KSNM 99).