Additional desynths (#1447)
* Update synth_recipes.sql

For each change, standards were: Desynth with FFXIAH listing the skill level as 255, BGWiki having a specific level, recipe and results match.

Recipes with a specific BGWiki skill level but a mismatch in yield or recepie had their skill level set, but remained commented out and had comment added at the end of the line.

---

For this round, I used the following criteria:

1: Item has a specific desynth recipe and yield listed on its BGWiki page (wihtout a level) AND Item has a synthesis recepie
2: Item was previously updated in level but not activated due to yield mismatch.

In cases where the item is craftable, but the desynth recepie had no level assigned to it, the level was set to match the synthesis level, and a comment was added in case future evidence proves that the level should be different.

In cases where the yield shown on FFXIAH/the current database were similar, but not exact matches, (More HQ result information on the BGWiki page primarily), the BGWiki data was used. This includes previously identified mismatched yields, and they have been activated as well.

---

ibm2431's note: All desynth changes were verified against the BGWiki item pages for the to-be-desynthed item on Nov 8 to Nov 11, 2020.