Update packet_system.cpp
This is something I shared with Teo privately a little while ago. Its method is based on how Zach locked down shady Ranged Attacks in the same file.

This prevents users from forcing Job Abilities through means that are NOT possible on retail. So this isn't blocking something like drive-by packet WS, which IS possible on retail (and super fun!). Previously, players could force their characters to use a job ability like Meditate, even if they were mounted on a chocobo.

This isn't a critical exploit, so just making a PR in broad daylight.