Add Against All Odds server midnight wait timer
Had an issue with this NPC elsewhere, discovered the timer wasn't in place for the JP midnight wait for re-acquiring another KI for the battle. This adds a timer on the initial CS, adds a check for it but also accounts for characters that didn't get it set previously on the NPC to re-acquire the KI, and clears it on the battlefield victory.

https://www.bg-wiki.com/ffxi/Against_All_Odds
https://ffxiclopedia.fandom.com/wiki/Against_All_Odds