Disables dropping of Moogle Slips containing items
Previously, players have been able to dispose of Porter Moogle Slips that contained stored items. On retail, this is prevented in order to protect players from losing those stored items. This implements a check for stored items and prevents it, giving players the retail accurate message.

Big thanks to @zach2good for doing most of the correction when my original stab at it wasn't producing the correct results.