Remove Vigil Weapons off Vajra Quadavs  (#98)
Remove Vigil Weapons off Vajra Quadavs

Metal discovered the Vajra Quadavs were dropping Vigil Weapons on the Canaria server. No recorded drops are to be found for them on a wiki or FFXIDB, so just blanked it out to dropid of 0. Did a quick search in mob_groups for ",906," and didn't turn up any more stray mob groups outside of Nyzul Isle with that droplist.