Fix Some ToAU Instance Blackscreens
Players completing ToAU 31 Shades of Vengeance battlefield and possibly players completing Assaults in Periqia (the later untested) were met with a black screen on CS 102 firing on instance completion. On local server, this was due to instance returning nil by this point in the process (instance complete-> play CS -> check for instance -- what instance?).

Simply removing these checks for the instance status on the victory animation removes the possibility of any of these issues. A user receiving that CS will be completing their instance anyway or will be a GM firing it intentionally.

Upon examination of Ilrusi Atol, I saw similar checks and feel it's best to remove them as well. Again, I don't believe it's necessary due to what has to happen for that event to fire.

Following the CS, the player is sent to the destination zone that interprets where they land from their prior zone in onZoneIn.

Also removed a few unnecessary parens per style guide.