Implement a basic 'Viscid Emission' mob TP skill
There is not a lot of info about this spell online. I based the implementation off of the existing `nepenthean_hum.lua`, which is also an AoE amnesia.

The difference being that this one is for a Cone (as reflected in the .sql).

Confirmed the animation by comparing with youtube videos from retail.