Implement a basic 'Severing Fang' mob TP skill
This implementation is largely based off of `razor_fang.lua`.

The skill is set to not trigger for targets behind the mob, since sub-zero smash should always be triggered in that case.

I also adjusted the shape and range in the .sql file. The range value was chosen based off of the one from `razor_fang`. And the shape was updated to be a cone since the shape is a "front arc".