Add basic implementation of `Aqua Blast` mob skill
There was not a whole lot available about this spell online, so I implemented it to behave largely like `aqua_breath.lua`. Also, for most of the knockback, I just put a placeholder value.

Most knockback spells appear to use `1` for the value in the SQL, so I put that as the placeholder in the SQL for this spell as well.

(This doc also mentioned as a TODO for this spell, as well: https://github.com/LandSandBoat/server/blob/base/documentation/Mobskills_Knockback_Values_TODO.xml#L12)

This change is done as part of https://github.com/LandSandBoat/server/issues/201