Remove duplicate `xi.zone.RESIDENTIAL_AREA` keys.
Lua tables take the last key in the initializer, so all of the earlier
keys were just being overwritten.

Also updated one comment that referred to the underlying value. (These
types of comments are brittle, as seen in this case, since they can
become stale. But I left it since it was pre-existing).