Rename duplicate keyitems to not conflict, add missing "lost article" no pun intended.
Not: of all these only the STAR_CRESTED_SUMMONS was used, and I d not know where the 2nd copy of it would be, we will need captures to confirm. Until then, everything will continue to use one KI as it always has which may or may not be correct. At least it works and from a player standpoint, there is no notable difference.