Disables dropping of Moogle Slips containing items  (#91)
* Update Ratihb.lua

Forgot a piece.

* Disables dropping of Moogle Slips containing items

Previously, players have been able to dispose of Porter Moogle Slips that contained stored items. On retail, this is prevented in order to protect players from losing those stored items. This implements a check for stored items and prevents it, giving players the retail accurate message.

Big thanks to @zach2good for doing most of the correction when my original stab at it wasn't producing the correct results.

* Update packet_system.cpp

* Delete Ratihb.lua

* Revert "Delete Ratihb.lua"

This reverts commit 7c5ee82982bbe748d607d9d309f945e2eb732fc9.

* Revert "Update Ratihb.lua"

This reverts commit 528209bf36bc43b8e5d055c8a3cacabd4d489904.

* Update packet_system.cpp

* requested adjustments

* Update item.cpp