Ashtu Talif missing onInstanceLoadFailed
Any zone used for instances like this should have this in place to direct the server where to spit you out at in the case of a failed instance load. See: https://github.com/LandSandBoat/server/blob/base/scripts/zones/Nyzul_Isle/Zone.lua

This is a for sure map crash pre-SOL, not sure of the results post-SOL. But an easy fix!