Add Kilusha Events and Shop
Kilusha is the Einherjar introductory NPC as well as your dealer in Smoldering Glasses and Therion Ichor items. NPC is all setup for later RoV KI adjustments once players have access to it. Requirements for the introduction CS has been personally verified on retail using multiple characters.

https://www.bg-wiki.com/ffxi/Kilusha
https://ffxiclopedia.fandom.com/wiki/Kilusha

Had discussed with @TeoTwawki how to handle the introduction CS and it was decided setting a CharVar on character creation was the best way. Current server operators will either need to add that CharVar to characters or it will have to be handled with dbtool somehow for them.

This requires items global update from PR #435 , currently pending merger.