Change mob_family_system table structure to include "super family" : Family = Korrigan, lycopodium etc. SuperFamily = Mandragora Ecosystem = Plantoid
Also renamed system to ecosystem to be more obvious.

ID were assigned on a first come first serve basis in spite of my temptation to try and organize this chaos and usage of superFamily in core will be in a followup commit. The strings can be anything and nothing stops you from mix and matching to create golem-mandragora-dragons if you wanted to.