1_1_The_Horutoto_Ruins_Experiment
What a fun mission to script as the first one in the Windy line -- NOT!

I fell into a rabbit hole of dead NPCs, stayed there too long, and had to hold my hands over my eyes and save that work for another day, which is really hard. They've made it this far and gods only know how long it will take for them to be brought to life! The goal is to just convert these Windurst missions to the new framework, though, so dialed back some of that endeavor and focused on the plan.

Major help from zach and clay, who provided some research on the various NPCs involved. Thanks to their assistance, this mission is in the most fleshed out state it's ever been in!

Reference links:
https://ffxiclopedia.fandom.com/wiki/The_Horutoto_Ruins_Experiment
https://www.bg-wiki.com/ffxi/Windurst_Mission_1-1

Co-Authored-By: Zach Toogood <zrtoogood@gmail.com>
Co-Authored-By: claywar <81713309+claywar@users.noreply.github.com>