Stop defender from removing berserk - fixes #868
The overwrite flag on these already says to allow them to overwrite themselves, but the type flag was telling them to erase effect ID 56, which was berserk..

On retail you can in fact use both and both stay, allowing you to temporarily compensate for one and click one back off as needed. They have exactly opposite effects and this was what the wiki page that says "canceled out" meant. This is why we're always insisting on in game verification and not just blindly trusting wiki pages!