[core] added code to prevent modules for re-caching
This is scene when loading mobs. For example, loadMOBList() loops
through all its mobs, each time it luautils::OnEntityLoad(PMob) once
the Mobs stats are set. This in turn called CacheLuaObjectFromFile()
on the mob even if it has already loaded it.

Ultimately, this creates a bug for people who load modules something
say the "Peryton" in Valley Of Sorrows. When it loads the Peryton
for the first time, it applies the override, and then marks that as
applied. However, once the second Peryton is loaded, it pulls directly
from the file and overwrites its cache value (which had been previously
overwritten by the module, essentially undoing that). However, when the
"TryApplyLuaModules" goes to run that override again, the applied flag
is already set to true so it does not try to load it again. Bother...