[Interaction] San d'oria 8-2 Add missionstatus checks to prevent skipping steps (#2504)
* Update Missionstatus Checks due to bug

In this mission, the player gets the mission from the gate guard and then is supposed to go for two cutscenes: The king, and Rahal, and thus updates their mission status to 2. Players are able to go directly to the KI portion of this mission and obtain the three KI which adjusts their mission status to 3 which then locks them permanently in that mission status. They should not be able to obtain this KI without talking to the king and Rahal first, therefore a mission status check of greater than 1 is required (utilized the greater than symbol due to the fact it is a +1 variable on each of the key item collections. so a flat number would not be appropriate.). 

The reproducibility of this error was tested on a base LSB server, and the correction was tested and fully functional.

* Update 8_2_Lightbringer.lua

* Remove spaces