Fix Dismounting sometimes causing players to become invisible/untargetable to others
This fix has a side effect of sending the associated dismount packet to all characters in the zone. I consider this low impact, but notable enough that I'd consider this fix temporary. A better solution would be to figure out where the packets the server is sending differ from retail.

Fixes #2514