Fix incorrect Droprates when Group Droprate totals are not 1000
If a mob's droplist includes a group where the group's total drop rates don't add up to 1000, the drop rates will be incorrect. There are two possibilities. The drop rates add up to less than 1000, resulting in a chance for the group to not drop even if the group should have a 100% drop chance. Or the drop rates add up to more than 1000, resulting in the drop rates being out of proportion and even making some items impossible to drop.

This fix rolls the droprate out of the total rate in the group, rather than always 1000.

Fixes #2527