Fix wyverns disappearing on certain types of zoning (#2537)
* Fix wyverns disappearing on certain types of zoning

Wyverns (and other pets retained through zoning) will not be respawned when leaving town (if you entered with it) and when zoning via the `luautils::setPos()` function, which includes NPC interactions and the `!zone` command.

* Switching to static_cast

Updating to the preferred static casting since we're touching this code anyway.

* DRY compliance

Cache the casted variable instead of repetitively casting.

* CPP Formatting compliance

* Jug pets should not persist

As per @WinterSolstice8, jug pets do not persist between zones on retail.

* Avatar should persist between zoning

As per testing by @WinterSolstice8, avatars should also be maintained through the different types of zoning.

* Avatar should persist through zoning

We already fixed the Avatar check for respawning the avatar, now we need to check for storing it. Also clean up JUG_PET at the same time, as they're not supposed to be saved between zones.