Fix Skillchains sometimes triggering before the 3 second window (#2534)
* Fix Skillchains sometimes triggering before the 3 second window

The code was not checking for the delay if it was a multi skillchain. That is, the 3rd or later WS in the skillchain.

* Cleanup nullptr check

This `nullptr` check is technically unnecessary, but it looks cleaner with it in.

* Formatting compliance

Failed a Clang check. I think it's this double empty line.

* CPP Formatting sanity check

Removing spaces from the empty line.