[misc] Update the wording for issue and PR templates
The previous wordings referred to 'this example' which was visible in
the text while editing, but not on the main page while viewing.
This makes it more obvious what happens if you don't fill in the boxes.

Also adds an obvious link to the Code of Conduct alongside the
Contributing Guide.

A couple of little grammar bits too.