Create Jailer_of_Prudence_2.lua (#130)
* Create Jailer_of_Prudence_2.lua

This PR is part of a three part fix to give functionality to our Jailer of Prudence module. We need to distinguish between the two Jailer of Prudence monsters for over rides to correctly apply. This PR creates a dummy file for over rides to be applied to.

* Update Jailer_of_Prudence_2.lua

Conform to requested CI