Convertes Bearclaw ENM x2 and Throne Room BCNM 60 to new framework
Converted ENMs Follow the white rabbit and when hell freezes over  to the new framework. Both logics for the mob behavior in their scripts need to be looked at as they are a bit wonky with their mechanics.

Converted Kindred Spirits to framework. work still needs to be done to pop the chest upon completion, but that is being worked on. Also it will current eat the orb instead of cracking it and returning to inventory.