Track the player's battlefield entered status with effect's subpower
The battlefield status effect subpower tracks if the player has entered the battlefield or not by using the first bit of the subpower. Everything after that is used for which battlefield area the player is in. This is done in order to know if the player needs to be ejected when logging in with a battlefield status effect.