[modules] Make ah_announcement.cpp more stable
The two queries that comprise the logic in here weren't guaranteed to be handling the same rows, so in a highly contested (multi process) environment messages may have been harmlessly sent to the wrong recipients.

Now, we store the seller id in LAST_INSERT_ID, which is guaranteed to be correct per-connection.

Each thread has its own connection, so no worries across threads.