Add `forceRepathInterval` option to `xi.follow`
This option will cause the follower to cancel its previous path early, to compensate for a faraway leader that might have moved long before the follower arrives.

Also, change default value of the `warpDistance` option to be disabled by default. And some minor refactoring for style and to reduce repetition.