Fix sunbreeze clustering issue
Fixed an issue where the sunbreeze event wasn't working as intended on servers that use clustering.
This PR forces entities to be handled and enabled on each specific's zone initialization.

style fix for PR

Fixed a style issue in the PR

Missing fix

Checks also required that the npcs be turned off if the check returns false