[login] Add more robust login error handling
Sometimes, a handler can throw an exception that will stop the io_context from continuing work.
This commit will allow the handlers to work after an exception of that type.

However, even truly fatal exceptions are ignored and we will want a list in the future of "acceptable" exceptions to ignore for socket exception reasons.