Remove manual deduction of HP from drain effects
A prior change causes the resistance calculations to deal the damage immediately where this did not happen before, meaning when we hit the defender with the hp removal, it'd take the damage twice.

There is likely now also an issue with non-hp drains also dealing damage, but that will be handled in a minor refactor already in progress.