add CDespawnState back to getCurrentAction()
This used to be here but was removed because in most places, we consider despawned mobs to also be "dead" and the next step is then considered waiting for respawn.

But if this is a mob that was not killed, but rather despawned by script in a zone suspended because its empty then it is not "dead" and it is not waiting to respawn either.

It won't transition until the 1st tick after the zone wakes (this is expected behavior).