Temp patch to clean action values of spikes that didn't proc
Technically, these should be the default values and then conditional branches change them instead of assuming damage before we even know if we have a proc.

However, it wasn't worth the effort when the whole thing is going to be eventually burned down to make way for fully scripted spikes.

This closes #4937

Co-Authored-By: TiberonKalkaz