Move Ashu Talif Captain logic into helper function, call in onMobEngaged
Added the captain's logic to a separate function and called it in his engaged and roam states. The mob would never roam as he would either spawn engaged to the player, or if the player is out of range, spawn engaged to Gessho. This caused the fight to not be able to be beaten unless you reraised.

Closes #4890
Closes #4790