[AMK, Lua] Mission 13 puzzle 1 - Beaucedine Glacier runaround (#5045)
* Add beauc glac puzzle npc logic

Structure for cutscenes handled by lonely evergreen and goblin grenadier, as well as the
structure of the pips.

* puzzle pips complete, map bug fixed

The cutscene event for the goblin grenadier requires two sets of args, first is for the hint and win/lose mechanic, the second
from the eventUpdate section is to relabel an pips on the map if you wanted to.  Setting the pip args to 0 makes it so the
map isn't changed accidentally