Era Quest Modules
Added individual quest modules that can be enabled for players to quickly access more "era" like behavior.

These modules are done on a per-quest-basis. This is to prevent modules from overlapping one another, and future proofing the contents.

Era Quest Timer Module

Creates a module for quests that had timers in their logic changed post release of seekers of Adoulin.
This PR includes 2 quests as an example for quests to come:
- "Your Crystal Ball"
- "Enveloped in Darkness"

It should be noted that this module should require that the quest have an interaction framework implemented to ensure its proper usage.

Create Subfolder

Created a subfolder for all era modules to live in.

module fix

Fixed incorrect file path for module